requirements.txt
aiohttp>=3.9,<4.0
python-dotenv>=1.0,<2.0
pynacl>=1.5,<2.0
